pyproject.toml
[tool.poetry]
name = "contextmesh"
version = "0.1.0"
description = "LLM-driven orchestration engine using OpenAPI extensions and Markdown playbooks"
readme = "README.md"
license = "MIT"
authors = ["ContextMesh Team"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{ include = "contextmesh", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
langchain = ">=1.1.0"
langchain-anthropic = ">=1.2.0"
langchain-core = ">=1.1.0"
pyyaml = ">=6.0"
pydantic = ">=2.0"
httpx = ">=0.27.0"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=4.1"
black = ">=24.0"
ruff = ">=0.1.0"
mypy = ">=1.8"

[tool.poetry.group.examples.dependencies]
fastapi = ">=0.115.0"
uvicorn = ">=0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/contextmesh"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
